requirements.txt
uvicorn
fastapi
psycopg2-binary==2.9
